seperate C implementation 2
