+++ packages/emnapi/index.js
@@ -7,7 +7,13 @@ const include = path.join(__dirname, 'include')
 const includeDir = path.relative(process.cwd(), include)
 const jsLibrary = path.join(__dirname, './dist/library_napi.js')
 const sources = [
+  path.join(__dirname, './src/js_native_api.c'),
+  path.join(__dirname, './src/node_api.c'),
   path.join(__dirname, './src/emnapi.c'),
+  path.join(__dirname, './src/async_cleanup_hook.c'),
+  path.join(__dirname, './src/async_context.c'),
+  path.join(__dirname, './src/async_work.c'),
+  path.join(__dirname, './src/threadsafe_function.c'),
   path.join(__dirname, './src/uv/uv-common.c'),
   path.join(__dirname, './src/uv/threadpool.c'),
   path.join(__dirname, './src/uv/unix/loop.c'),
